feat: Global Entry List (#291)
* feat: Global Entry List

* fix

* move admin list to global entry list

* fix

* Refactor server configuration components and improve global entry/ban list functionality

- Updated entry.vue and entrylist.vue to enhance data handling and UI consistency.
- Improved configuration.vue for better tab management and layout.
- Enhanced globalbanlist.vue and globalentrylist.vue for better user experience and error handling.
- Cleaned up index.js and status.vue for improved readability and maintainability.
- Refactored store.js for consistent code style.

* chore: update changelog for version 1.27.0 with detailed entries

+++ public/src/components/server.vue
@@ -1,29 +1,31 @@
 <template>
-    <div class="server">
-        <div>
+    <div class="server" v-bind:class="{running: formattedServerClientCount > 0}">
+        <div class="content">
             <div class="name">
                 {{server.name}}
-                <span v-if="is_ro">
-                    <i class="fas fa-tv" v-if="server.pid" v-on:click="live" :title="$t('view_live')"></i>
-                </span>
-                <span v-if="!ro">
-                    <i class="fas fa-cog" v-on:click="edit" :title="$t('change_config')"></i>
-                    <i class="fas fa-terminal" v-on:click="logs" :title="$t('view_logs')"></i>
-                    <i class="fas fa-copy" v-on:click="copyConfig" v-if="is_admin" :title="$t('copy_config')"></i>
-                    <i class="fas fa-file-download" v-on:click="exportConfig" :title="$t('export_config')"></i>
-                    <i class="fas fa-trash" v-on:click="deleteServer" v-if="is_admin" :title="$t('delete_server')"></i>
-                </span>
+                <div class="actions">
+                    <span v-if="is_ro">
+                        <i class="fas fa-tv" v-if="server.pid" v-on:click="live" :title="$t('view_live')"></i>
+                    </span>
+                    <span v-if="!ro">
+                        <i class="fas fa-cog" v-on:click="edit" :title="$t('change_config')"></i>
+                        <i class="fas fa-terminal" v-on:click="logs" :title="$t('view_logs')"></i>
+                        <i class="fas fa-copy" v-on:click="copyConfig" v-if="is_admin" :title="$t('copy_config')"></i>
+                        <i class="fas fa-file-download" v-on:click="exportConfig" :title="$t('export_config')"></i>
+                        <i class="fas fa-trash" v-on:click="deleteServer" v-if="is_admin" :title="$t('delete_server')"></i>
+                    </span>
+                </div>
             </div>
             <div class="info">
-                <span v-if="server.pid">PID: {{server.pid}}</span>
-                UDP: {{server.udpPort}} &bull;
-                TCP: {{server.tcpPort}} &bull;
-                {{$t("track")}}: {{server.track}}
-                <span v-if="!ro">&bull; {{$t("configuration_directory")}}: {{server.id}}</span>
+                <span v-if="server.pid"><b>PID:</b> {{server.pid}} &nbsp;&bull;&nbsp;</span>
+                <b>UDP:</b> {{server.udpPort}} &nbsp;&bull;&nbsp;
+                <b>TCP:</b> {{server.tcpPort}} &nbsp;&bull;&nbsp;
+                <b>{{$t("track")}}:</b> {{server.track}}
+                <span v-if="!ro">&nbsp;&bull;&nbsp; <b>{{$t("configuration_directory")}}:</b> {{server.id}}</span>
             </div>
             <div class="info state" v-if="server.pid">
-                <b>{{$t("state")}}: </b>{{$t(server.serverState)}} &bull;
-                <b>{{$t("number_of_drivers")}}: </b>{{formattedServerClientCount}} &bull;
+                <b>{{$t("state")}}: </b>{{$t(server.serverState)}} &nbsp;&bull;&nbsp;
+                <b>{{$t("number_of_drivers")}}: </b>{{formattedServerClientCount}} &nbsp;&bull;&nbsp;
                 <b>{{$t("session")}}: </b>
                 <span v-if="server.sessionType">{{server.sessionType}} ({{server.sessionPhase}}) - {{server.sessionRemaining}} min(s)</span>
                 <span v-else>{{$t('not_detected')}}</span>
@@ -37,12 +39,26 @@
 </template>
 
 <style>
+.content {
+    width: 100%;
+}
+
 .state {
     margin-top: 10px;
 }
 
 .state b {
     color: #505050;
+}
+
+.running {
+    background-color: #1d2331;
+}
+
+.actions {
+    display: inline;
+    float: right;
+    margin-right: 30px;
 }
 </style>
 
@@ -84,6 +100,10 @@ export default {
             link.remove();
         },
         deleteServer() {
+            if (!window.confirm(this.$t("confirm_delete_server"))) {
+                return;
+            }
+
             axios.delete(`/api/instance/${this.server.id}`)
             .then(() => {
                 this.$emit("deleted");
@@ -125,6 +145,7 @@ export default {
         "copy_config": "Copy config",
         "export_config": "Export config",
         "delete_server": "Delete server",
+        "confirm_delete_server": "Do you really want to delete this server?",
         "copy_server_error": "Error copying server configuration.",
         "delete_server_error": "Error deleting server configuration.",
         "start_server_error": "Error starting server, please check the logs. ERROR: {error}",
